Fix readme.md (#47)
* feat(DatabaseResolver): add posgresql support

Added postgresql support. Now the SQLAlchemyDatabaseResolver
will connect to a postgresql database instance.

* feat(DatabaseResolver): add psycopg2 dependency

Added psycopg2 to requirements.
This library is used by the SQLAlchemyDatabaseResolver
to connect to a postgresql database instance.

* feat(DatabaseResolver): add testing.postgresql dependency

Added testing.postgresql to requirements.
This library is used for making a test
postgresql database.

* feat(DatabaseResolver): add linebreaks for pep8 format

Added linebreaks. Now the file is conform pep8 style

* feat(DatabaseResolver): add flake8 dependency

Added flake8. Flake8 is used for formatting and checking

* ref(DatabaseResolver): changed imports

Changed Context import to AlgorithmContext.

* ref(DatabaseResolver): add None parameter for Context

Added the default None reference to AlgorithmContext

* feature(DatabaseResolver): add State behaviour check

Added check for State instance and check
for state methods, if class is not a subclass or a State

* ref(DatabaseResolver): Changed Context to AlgorithmContext

Refactored import from Context to AlgorithmContext

* feature(DatabaseResolver): Added AlgorithmContext tests

Added tests AlgorithmContext

* feature(DatabaseResolver): Added StateValidators tests

Added tests StateValidators

* feature(DatabaseResolver): Added DataProvidingState tests

Added tests for DataProvidingState

* ref(DatabaseResolver): Change Context to AlgorithmContext

Change Context to AlgorithmContext import

* feat(DatabaseResolver): Add psycopg2 dependency

* ref(DatabaseResolver): Change Context import

Change Context import to AlgorithmContext

* fix(Readme): fix link to build

Fixed link to travis-ci build. The link was pointing to
a wrong repo.

Co-authored-by: coding-kitties <59013545+coding-kitties@users.noreply.github.com>